% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchStaisticsWithPvalue.R
\name{MatchStaisticsWithPvalue}
\alias{MatchStaisticsWithPvalue}
\title{JunctionSeq:::MatchStaisticsWithPvalue(Re.example.statistics, re.example.gene.based)}
\usage{
MatchStaisticsWithPvalue(Re.example.statistics, re.example.gene.based)
}
\description{
JunctionSeq:::MatchStaisticsWithPvalue(Re.example.statistics, re.example.gene.based)
}
