\docType{class}
\name{JunctionSeqCountSet-class}
\alias{JunctionSeqCountSet}
\alias{JunctionSeqCountSet-class}
% Constructors:
\alias{newJunctionSeqCountSet}

\title{JunctionSeqCountSet object and constructors}

\description{
  A \code{JunctionSeqCountSet} is a container class that 
  contains all information pertaining to a JunctionSeq analysis
  and dataset.
}

\section{Constructor}{
  \describe{
    \item{}{

      \code{newJunctionSeqCountSet( countData, 
                              geneCountData, 
                              design, 
                              geneIDs, 
                              countbinIDs, 
                              featureIntervals=NULL, 
                              transcripts=NULL)}:
      Creates a new JunctionSeqCountSet
      
      \describe{
        \item{\code{countData}}{
    A matrix of junction-level count data of non-negative integer values. 
    The rows correspond to counts for each splice-junction counting bin, 
    the columns correspond to samples. Note that biological 
    replicates should each get their own column, while the 
    counts of technical replicates (i.e., several sequencing 
    runs/lanes from the same sample) should be summed up 
    into a single column.
        }
        \item{\code{geneCountData}}{
    A matrix of gene-level count data of non-negative integer values. 
    The rows correspond to counts for each gene, 
    the columns correspond to samples. Note that biological 
    replicates should each get their own column, while the 
    counts of technical replicates (i.e., several sequencing 
    runs/lanes from the same sample) should be summed up 
    into a single column. Must have the same dimensions as countData.
        }
        \item{\code{design}}{
    A data frame consisting of all factors to be included in the analysis.
    All columns should be factors. Each column should represent a different variable,
    each row should represent a different sample. The number of rows must equal the
    number of columns in geneCountData and countData.
        }
        \item{\code{geneIDs}}{
    A character vector of gene indentifiers for each splice junction. The length must
    equal the number of rows in countData.
        }
        \item{\code{countbinIDs}}{
    A character vector of splice-junction-locus indentifiers for each splice junction. The length must
    equal the number of rows in countData.
        }
        \item{\code{featureIntervals}}{
    Optional. A data.frame with 4 columns: "chr", "start", "end", and "strand". chr and strand should 
    be character vectors or factors, start and end must be integers.  
        }
        \item{\code{transcripts}}{
    Optional. Character vector listing the transcripts that each splice junction belongs to. Some junctions
    may belong to more than one transcripts. In this case, transcripts should be separated with the "+" character.
        }
      }
      
      This constructor function SHOULD NOT BE USED in normal operation. Instead you should use the \code{\link{readJunctionSeqCounts}} function, which returns a new \code{JunctionSeqCountSet}.
    
    }
  }
}

\details{
  Slots:
  \describe{
    \item{\code{designColumns }}{  
      A character vector with the column names in the \code{design} data.frame.
    }
    \item{\code{dispFitCoefs}}{  
      The dispersion fit coefficients.
    }
    \item{\code{fittedMu}}{  
      Fitted mu values generated by DESeq2 code.
    }
    \item{\code{dispFunctionType}}{  
      A list of various variables defining the dispersion functon used.
    }
    \item{\code{dispFunction}}{  
      A function that converts a base mean to a fitted dispersion based on all included count bins.
    }
    \item{\code{dispFunctionJct}}{  
      A function that converts a base mean to a fitted dispersion based only on the splice junction bins.
    }
    \item{\code{dispFunctionExon }}{  
      A function that converts a base mean to a fitted dispersion based only on the exon bins.
    }
    \item{\code{formulas}}{  
      A list of formulas used.
    }
    \item{\code{annotationFile}}{  
      The annotation file.
    }
    \item{\code{geneCountData}}{  
      A matrix of the gene-level counts
    }
    \item{\code{countVectors}}{  
      A matrix of the count vectors.
    }
    \item{\code{altSizeFactors}}{  
      (Not currently used)
    }
    \item{\code{plottingEstimates}}{  
      A list of fitted estimates, for plotting.
    }
    \item{\code{plottingEstimatesVST }}{  
      (Not currently used)
    }
    \item{\code{geneLevelPlottingEstimates }}{  
      A list of gene-level fitted estimates, for plotting.
    }
    \item{\code{modelFitForHypothesisTest }}{  
      (Not currently used)
    }
    \item{\code{modelFitForEffectSize }}{  
      (Not currently used)
    }
    \item{\code{flatGffData}}{  
      A \code{data.frame} representation of the flattened gff annotation for each countbin.
    }
    \item{\code{flatGffGeneData}}{  
      A \code{data.frame} representation of the flattened gff annotation for each gene.
    }
    \item{\code{analysisType}}{  
      The type of analysis. Character string.
    }
    \item{\code{DESeqDataSet}}{  
      The specially-constructed \code{DESeqDataSet}, to be passed to the internally-loaded DESeq2 code.
    }
  }
}

\seealso{
  \code{\link{readJunctionSeqCounts}}
}