\name{readJunctionSeqCounts}
\docType{methods}
\alias{readJunctionSeqCounts}
\title{
   Read junctionSeq count files
}
\description{
   This function reads count data (usually produced by QoRTs) and 
   compiles them into a \code{JunctionSeqCountSet}.
   
   For most purposes this function should not be needed, but might 
   be useful to advanced users performing non-standard analyses. 
   This and all other standard analysis functions can be run 
   automatically via the \code{\link{runJunctionSeqAnalyses}} 
   function.
}
\usage{
readJunctionSeqCounts(countfiles, 
                      countdata,
                      samplenames,
                      design, 
                      flat.gff.file, 
                      use.splice.sites = TRUE, 
                      use.novel.splice.sites = TRUE, 
                      use.exons = FALSE, 
                      verbose = TRUE)
}

\arguments{
  \item{countfiles}{
    Character vector. The filenames of the count files generated by QoRTs. The counts must all
    be generated using equivalent QoRTs parameters. The strandedness must be the same, as well
    as the inclusion of novel junctions.
  }
  \item{countdata}{
    List. An alternative parameterization. Instead of supplying count files using the 
    countfiles parameter, you can pass a list of data frames, one for each sample. 
    Each data frame should contain two columns: the first should be the feature id 
    and the second should be the counts. This list must have the same length as 
    the samplenames parameter.
  }
  
  \item{samplenames}{
    Character vector. A vector of full sample names, in the same order as the countfiles parameter. 
  }
  \item{design}{
    A data frame containing the condition variable and all desired covariates. All variables should be
    factors.
  }
  \item{flat.gff.file}{
    Character string. The filename of the "flat" gff annotation file. Can be gzip-compressed. 
    This "flat" gff file must be produced by the QoRTs jar utility using the makeFlatGtf or 
    mergeNovelSplices functions (depending on whether inclusion of novel splice junctions is desired).
    
    NOTE: This option is technically optional, but strongly recommended. If it is not included, then 
    attempts to plot the results will crash unless (non-default) options are used to deactivate the 
    plotting of genomic coordinates and transcript information
  }
  \item{use.splice.sites}{
    FOR ADVANCED USERS ONLY: Logical value. If \code{TRUE}, splice junctions will be included in the analyses.
  }
  \item{use.novel.splice.sites}{
    Logical value. If \code{TRUE}, splice junctions that are marked as novel in the gff annotation will be included in the analysis.
    Note that this is does nothing if there are no novel splice junctions in the count tables. 
  }
  \item{use.exons}{
    FOR ADVANCED USERS ONLY: Logical value. If \code{TRUE}, exon segments will be included in the analyses.
  }
  \item{verbose}{if TRUE, send debugging and progress messages to the console / stdout.}
}
\value{
  A JunctionSeqCountSet.
}